+++ NEWS.md
@@ -10,3 +10,8 @@
 
 * Simplify make_grid function to improve the splits of the divisions and strata
 * CRAN request: change doi in DESCRIPTION to prefered format
+
+# SimSurvey 0.1.3.009
+
+* Add a "bezier" method to make_grid
+* Allow a vector of age-specific parameters to be supplied to sim_parabola plus add some options for defining a more asymetric parabola